Made adjustments based on various mods I've been working on
Commented out example code, as I tend to forget to remove the config options for them.
Return the Coroutine when it's started from Plugin.Instance, allowing for more control over the coroutines. (Used in NijiroScoring mod)
Changed LogType.Debug to use LogInfo instead of LogDebug, but only when built in Debug configuration, making it easier to use.